api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic==2.11.4
